allows to build on Ubuntu18.04

+++ utils/dbt3/dss.h
@@ -486,10 +486,10 @@ int dbg_print(int dt, FILE *tgt, void *data, int len, int eol);
 #endif /* PGSQL */
 #ifdef MDY_DATE
 #define  PR_DATE(tgt, yr, mn, dy)	\
-   sprintf(tgt, "%02d-%02d-19%02d", mn, dy, yr)
+	sprintf(tgt, "%02d-%02d-19%02d", (int)(mn), (int)(dy), (int)(yr))
 #else
 #define  PR_DATE(tgt, yr, mn, dy)	\
-sprintf(tgt, "19%02ld-%02ld-%02ld", yr, mn, dy)
+	sprintf(tgt, "19%02d-%02d-%02d", (int)(yr), (int)(mn), (int)(dy))
 #endif /* DATE_FORMAT */
 
 /*
